Display random coins as bookmarks, just USD value.
Detailed changes:
- Add layout for the Bookmark card.
- Create BookmarksFragment layout, complete with SwipeRefreshLayout,
RecyclerView & Toolbar.
- Connect View & Presenter s.t. a list of randomly selected coins
are displayed as bookmarks.
- Use UserSettings in various places to select the primary currency
in which a coin's value will be displayed(besides bitcoin)
- Create DBCryptoCoinSinglePriceData object which retrievs from the
DB a DbCryptoCoin with price data in just one currency. Required
since the DbCryptoCoin uses the @Relation annotation for the price
data, annotation which does not currently allow to specify select
conditions.

Type: New Feature